+++ src/main/java/com/gokimpark/instaclone/domain/post/PostRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer> {
 
     List<Post> findAllByUser(User user);
     void deleteAllByUser(User user);
